refactor(codec): collapse *Transform into *Handler::redact; merge transform/ into handler/
- Fold each *Transform trait into the matching *Handler as a provided
  default method redact(Redactions<L, R>) that loops redact_at in
  insertion order. AudioHandler::redact overrides the default to sort
  right-to-left by time_span.start_us so AudioOutput::Remove doesn't
  shift later sample indices. One trait per modality instead of two;
  no separate blanket-impl extension trait.
- Move every transform/ contents into handler/:
  - *Redaction / *Output structs into handler/{text,image,audio,tabular}/
    next to the redact_at hook they feed.
  - Redactions<S, R>, ConflictPolicy, InsertError into handler/ as
    the cross-modality batching primitives.
  - Mergeable re-exported from handler/ (still defined in ontology).
- Delete transform/ entirely; nothing in the engine or codec imports
  from it anymore. Engine and downstream consumers now go through
  nvisy_codec::handler::{*Redaction, Redactions, ConflictPolicy, ...}.
