+++ crates/nvisy-codec/src/document/mod.rs
@@ -30,9 +30,8 @@ use crate::handler::{DocxLoader, DocxParams};
 use crate::handler::{HtmlLoader, HtmlParams};
 #[cfg(feature = "pdf")]
 use crate::handler::{PdfLoader, PdfParams};
-use crate::transform::{
-    AudioRedaction, AudioTransform, ImageRedaction, ImageTransform, Redactions, TabularRedaction,
-    TabularTransform, TextRedaction, TextTransform,
+use crate::handler::{
+    AudioRedaction, ImageRedaction, Redactions, TabularRedaction, TextRedaction,
 };
 
 /// A fully type-erased document that can hold any supported format.
@@ -171,8 +170,8 @@ impl ContentHandle {
         redactions: Redactions<TextLocation, TextRedaction>,
     ) -> Result<(), Error> {
         match self {
-            Self::Text(h) => TextTransform::redact(h, redactions).await,
-            Self::Rich(h) => TextTransform::redact(h, redactions).await,
+            Self::Text(h) => TextHandler::redact(h, redactions).await,
+            Self::Rich(h) => TextHandler::redact(h, redactions).await,
             Self::Tabular(_) | Self::Image(_) | Self::Audio(_) => Ok(()),
         }
     }
@@ -183,7 +182,7 @@ impl ContentHandle {
         redactions: Redactions<TabularLocation, TabularRedaction>,
     ) -> Result<(), Error> {
         match self {
-            Self::Tabular(h) => TabularTransform::redact(h, redactions).await,
+            Self::Tabular(h) => h.redact(redactions).await,
             _ => Ok(()),
         }
     }
@@ -194,8 +193,8 @@ impl ContentHandle {
         redactions: Redactions<ImageLocation, ImageRedaction>,
     ) -> Result<(), Error> {
         match self {
-            Self::Image(h) => ImageTransform::redact(h, redactions).await,
-            Self::Rich(h) => ImageTransform::redact(h, redactions).await,
+            Self::Image(h) => ImageHandler::redact(h, redactions).await,
+            Self::Rich(h) => ImageHandler::redact(h, redactions).await,
             Self::Text(_) | Self::Tabular(_) | Self::Audio(_) => Ok(()),
         }
     }
@@ -206,7 +205,7 @@ impl ContentHandle {
         redactions: Redactions<AudioLocation, AudioRedaction>,
     ) -> Result<(), Error> {
         match self {
-            Self::Audio(h) => AudioTransform::redact(h, redactions).await,
+            Self::Audio(h) => h.redact(redactions).await,
             _ => Ok(()),
         }
     }
