+++ crates/nvisy-codec/src/handler/audio/mod.rs
@@ -1,15 +1,18 @@
 //! Audio format handlers and loaders.
 
+use std::cmp::Reverse;
+
 use nvisy_core::Error;
 use nvisy_ontology::entity::AudioLocation;
 
 use super::Handler;
 use crate::document::LocationStream;
-use crate::transform::AudioRedaction;
+use crate::handler::Redactions;
 
 mod apply;
 mod audio_data;
 mod audio_handler;
+mod instruction;
 mod mp3_handler;
 mod mp3_loader;
 mod wav_handler;
@@ -17,6 +20,7 @@ mod wav_loader;
 
 pub(crate) use self::apply::apply_audio_redaction;
 pub use self::audio_data::AudioData;
+pub use self::instruction::{AudioOutput, AudioRedaction};
 pub use self::audio_handler::BoxedAudioHandler;
 pub use self::mp3_handler::Mp3Handler;
 pub use self::mp3_loader::{Mp3Loader, Mp3Params};
@@ -31,12 +35,16 @@ pub use self::wav_loader::{WavLoader, WavParams};
 ///   location.
 /// - [`redact_at`]: apply a single redaction at a single time range.
 ///
-/// Batched redaction lives on the blanket-impl [`AudioTransform::redact`].
+/// Batched redaction is provided by [`redact`], which overrides the
+/// default loop ordering to apply later time spans first — an
+/// [`AudioOutput::Remove`] shrinks the buffer and shifts every later
+/// sample index, so right-to-left order keeps earlier indices valid.
 ///
 /// [`locations`]: AudioHandler::locations
 /// [`read`]: AudioHandler::read
 /// [`redact_at`]: AudioHandler::redact_at
-/// [`AudioTransform::redact`]: crate::transform::AudioTransform::redact
+/// [`redact`]: AudioHandler::redact
+/// [`AudioOutput::Remove`]: crate::handler::AudioOutput::Remove
 #[async_trait::async_trait]
 pub trait AudioHandler: Handler {
     /// Async stream of [`AudioLocation`]s for this document, each
@@ -57,4 +65,19 @@ pub trait AudioHandler: Handler {
         location: &AudioLocation,
         redaction: AudioRedaction,
     ) -> Result<(), Error>;
+
+    /// Apply every `(location, redaction)` pair in `redactions` to the
+    /// handler, sorted right-to-left by `time_span.start_us`. The first
+    /// error aborts the batch.
+    async fn redact(
+        &mut self,
+        redactions: Redactions<AudioLocation, AudioRedaction>,
+    ) -> Result<(), Error> {
+        let mut items: Vec<_> = redactions.into_iter().collect();
+        items.sort_by_key(|(loc, _)| Reverse(loc.time_span.start_us));
+        for (location, redaction) in items {
+            self.redact_at(&location, redaction).await?;
+        }
+        Ok(())
+    }
 }
