+++ crates/nvisy-codec/src/handler/tabular/mod.rs
@@ -18,17 +18,19 @@ use nvisy_ontology::entity::TabularLocation;
 use super::Handler;
 use crate::document::LocationStream;
 use crate::handler::TextData;
-use crate::transform::TabularRedaction;
+use crate::handler::Redactions;
 
 mod apply;
 mod csv_handler;
 mod csv_loader;
+mod instruction;
 mod tabular_handler;
 mod xlsx_handler;
 mod xlsx_loader;
 
 pub(crate) use self::apply::apply_tabular_redaction;
 pub use self::csv_handler::{CsvData, CsvHandler};
+pub use self::instruction::TabularRedaction;
 pub use self::csv_loader::{CsvLoader, CsvParams};
 pub use self::tabular_handler::BoxedTabularHandler;
 pub use self::xlsx_handler::XlsxHandler;
@@ -42,12 +44,13 @@ pub use self::xlsx_loader::{XlsxLoader, XlsxParams};
 /// - [`read`]: fetch a cell's value as [`TextData`].
 /// - [`redact_at`]: apply a single redaction to a single cell.
 ///
-/// Batched redaction lives on the blanket-impl [`TabularTransform::redact`].
+/// Batched redaction is provided by [`redact`], which loops
+/// [`redact_at`] in insertion order.
 ///
 /// [`locations`]: TabularHandler::locations
 /// [`read`]: TabularHandler::read
 /// [`redact_at`]: TabularHandler::redact_at
-/// [`TabularTransform::redact`]: crate::transform::TabularTransform::redact
+/// [`redact`]: TabularHandler::redact
 #[async_trait::async_trait]
 pub trait TabularHandler: Handler {
     /// Async stream of [`TabularLocation`]s for this document, each
@@ -69,4 +72,22 @@ pub trait TabularHandler: Handler {
         location: &TabularLocation,
         redaction: TabularRedaction,
     ) -> Result<(), Error>;
+
+    /// Apply every `(location, redaction)` pair in `redactions` to the
+    /// handler in insertion order. The first error aborts the batch.
+    ///
+    /// The default loops [`redact_at`] in [`Redactions`] insertion
+    /// order; handlers with ordering constraints override this
+    /// default.
+    ///
+    /// [`redact_at`]: TabularHandler::redact_at
+    async fn redact(
+        &mut self,
+        redactions: Redactions<TabularLocation, TabularRedaction>,
+    ) -> Result<(), Error> {
+        for (location, redaction) in redactions {
+            self.redact_at(&location, redaction).await?;
+        }
+        Ok(())
+    }
 }
