+++ crates/nvisy-codec/src/handler/tabular/tabular_handler.rs
@@ -10,7 +10,7 @@ use nvisy_ontology::entity::TabularLocation;
 use super::TabularHandler;
 use crate::document::LocationStream;
 use crate::handler::{CsvHandler, Handler, TextData, XlsxHandler};
-use crate::transform::TabularRedaction;
+use crate::handler::TabularRedaction;
 
 /// A type-erased tabular handler backed by a boxed trait object.
 pub struct BoxedTabularHandler(Box<dyn TabularHandler>);
