+++ crates/nvisy-codec/src/handler/text/mod.rs
@@ -5,13 +5,14 @@ use nvisy_ontology::entity::TextLocation;
 
 use super::Handler;
 use crate::document::LocationStream;
-use crate::transform::TextRedaction;
+use crate::handler::Redactions;
 
 mod apply;
 #[cfg(feature = "html")]
 mod html_handler;
 #[cfg(feature = "html")]
 mod html_loader;
+mod instruction;
 mod json_handler;
 mod json_loader;
 mod markdown_loader;
@@ -21,6 +22,7 @@ mod txt_handler;
 mod txt_loader;
 
 pub(crate) use self::apply::apply_text_redaction;
+pub use self::instruction::{TextOutput, TextRedaction};
 
 #[cfg(feature = "html")]
 pub use self::html_handler::{HtmlData, HtmlHandler};
@@ -66,13 +68,32 @@ pub trait TextHandler: Handler {
 
     /// Apply a single redaction at the given location, mutating in
     /// place. Implementations need not handle iteration or overlap —
-    /// the blanket [`TextTransform::redact`] feeds one
-    /// `(location, redaction)` pair at a time.
+    /// the provided [`redact`] feeds one `(location, redaction)` pair
+    /// at a time.
     ///
-    /// [`TextTransform::redact`]: crate::transform::TextTransform::redact
+    /// [`redact`]: TextHandler::redact
     async fn redact_at(
         &mut self,
         location: &TextLocation,
         redaction: TextRedaction,
     ) -> Result<(), Error>;
+
+    /// Apply every `(location, redaction)` pair in `redactions` to the
+    /// handler in insertion order. The first error aborts the batch.
+    ///
+    /// The default loops [`redact_at`] in [`Redactions`] insertion
+    /// order; handlers with ordering constraints (see
+    /// [`AudioHandler::redact`]) override this default.
+    ///
+    /// [`redact_at`]: TextHandler::redact_at
+    /// [`AudioHandler::redact`]: crate::handler::AudioHandler::redact
+    async fn redact(
+        &mut self,
+        redactions: Redactions<TextLocation, TextRedaction>,
+    ) -> Result<(), Error> {
+        for (location, redaction) in redactions {
+            self.redact_at(&location, redaction).await?;
+        }
+        Ok(())
+    }
 }
