+++ crates/nvisy-codec/src/handler/text/txt_handler.rs
@@ -18,8 +18,7 @@ use nvisy_ontology::entity::TextLocation;
 use crate::document::{Located, LocationStream};
 use crate::handler::text::TextData;
 use crate::handler::{Handler, TextHandler};
-use super::apply_text_redaction;
-use crate::transform::TextRedaction;
+use super::{TextRedaction, apply_text_redaction};
 
 const TARGET: &str = "txt-handler";
 
@@ -168,7 +167,8 @@ mod tests {
     use nvisy_core::Error;
 
     use super::*;
-    use crate::transform::{ConflictPolicy, Redactions, TextOutput, TextTransform};
+    use crate::handler::TextHandler;
+    use crate::handler::{ConflictPolicy, Redactions, TextOutput};
 
     fn handler(text: &str) -> TxtHandler {
         let trailing_newline = text.ends_with('\n');
