--- crates/nvisy-codec/src/transform/audio/mod.rs
@@ -1,7 +1,0 @@
-//! Audio redaction primitives.
-
-mod instruction;
-mod transform;
-
-pub use self::instruction::{AudioOutput, AudioRedaction};
-pub use self::transform::AudioTransform;
