--- crates/nvisy-codec/src/transform/image/mod.rs
@@ -1,7 +1,0 @@
-//! Image redaction primitives.
-
-mod instruction;
-mod transform;
-
-pub use self::instruction::{ImageOutput, ImageRedaction};
-pub use self::transform::ImageTransform;
