--- crates/nvisy-codec/src/transform/tabular/mod.rs
@@ -1,7 +1,0 @@
-//! Tabular redaction primitives.
-
-mod instruction;
-mod transform;
-
-pub use self::instruction::TabularRedaction;
-pub use self::transform::TabularTransform;
