--- crates/nvisy-codec/src/transform/text/mod.rs
@@ -1,7 +1,0 @@
-//! Text redaction primitives.
-
-mod instruction;
-mod transform;
-
-pub use self::instruction::{TextOutput, TextRedaction};
-pub use self::transform::TextTransform;
